backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.17.1
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
cachetools==6.2.1
    # via premailer
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
chardet==5.2.0
    # via emails
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cssselect==1.3.0
    # via premailer
cssutils==2.11.1
    # via
    #   emails
    #   premailer
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic
emails==0.6
    # via -r requirements.in
fastapi[standard]==0.121.1
    # via
    #   -r requirements.in
    #   sentry-sdk
fastapi-cli[standard]==0.0.16
    # via fastapi
fastapi-cloud-cli==0.3.1
    # via fastapi-cli
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   fastapi
lxml==6.0.2
    # via
    #   emails
    #   premailer
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via cssutils
passlib[bcrypt]==1.7.4
    # via -r requirements.in
premailer==3.10.0
    # via emails
psycopg[binary]==3.2.12
    # via -r requirements.in
psycopg-binary==3.2.12
    # via psycopg
pydantic[email]==2.12.4
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via emails
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   -r requirements.in
    #   fastapi
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via
    #   emails
    #   premailer
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk[fastapi]==2.44.0
    # via
    #   -r requirements.in
    #   fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.27
    # via -r requirements.in
starlette==0.49.3
    # via fastapi
tenacity==8.5.0
    # via -r requirements.in
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
pandas==2.2.3
    # added for CSV upload/processing
shap>=0.41.0
    # for explainability (SHAP values)
scikit-learn==1.6.1
    # ML model dependencies (pinned to match pickled model version)
slowapi>=0.1.9
    # rate limiting for API endpoints
numpy>=1.24.0
    # numerical computations
matplotlib>=3.7.0
    # plotting for SHAP visualizations
joblib>=1.3.0
    # model serialization

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.0.0
testcontainers>=3.7.0
